Trait centrifuge_chain::rpc::anchors::AnchorApiServer

pub trait AnchorApiServer<IdHash, BlockHash>: Sized + Send + Sync + 'static {
    fn get_anchor_by_id(
        &self, 
        id: IdHash, 
        at: Option<BlockHash>
    ) -> RpcResult<AnchorData<IdHash, BlockNumber>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        IdHash: Send + Sync + 'static + DeserializeOwned + Serialize,
        BlockHash: Send + Sync + 'static + DeserializeOwned,
    { ... }
}

Server trait implementation for the AnchorApi RPC API.

Required Methods

fn get_anchor_by_id(
&self,
id: IdHash,
at: Option<BlockHash>
) -> RpcResult<AnchorData<IdHash, BlockNumber>>

Returns an anchor given an anchor id from the runtime storage

Provided Methods

fn into_rpc(self) -> RpcModule<Self> where
IdHash: Send + Sync + 'static + DeserializeOwned + Serialize,
BlockHash: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.