Function centrifuge_chain::service::evm::start_node_impl

pub(crate) async fn start_node_impl<RuntimeApi, Executor, RB, BIQ, BIC>(
    parachain_config: Configuration,
    polkadot_config: Configuration,
    eth_config: EthConfiguration,
    collator_options: CollatorOptions,
    id: ParaId,
    first_evm_block: BlockNumber,
    rpc_ext_builder: RB,
    build_import_queue: BIQ,
    build_consensus: BIC
) -> Result<(TaskManager, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>)>where
    RuntimeApi: ConstructRuntimeApi<Block, TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: TaggedTransactionQueue<Block> + Metadata<Block> + SessionKeys<Block> + ApiExt<Block, StateBackend = StateBackendFor<TFullBackend<Block>, Block>> + OffchainWorkerApi<Block> + BlockBuilder<Block> + CollectCollationInfo<Block> + EthereumRuntimeRPCApi<Block> + ConvertTransactionRuntimeApi<Block>,
    StateBackendFor<TFullBackend<Block>, Block>: StateBackend<BlakeTwo256>,
    Executor: NativeExecutionDispatch + 'static,
    RB: Fn(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, Arc<FullPool<Block, TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>>, DenyUnsafe, SubscriptionTaskExecutor, Arc<NetworkService<Block, Hash>>, Arc<FrontierBackend<Block>>, FilterPool, FeeHistoryCache, Arc<OverrideHandle<Block>>, Arc<EthBlockDataCacheTask<Block>>) -> Result<RpcModule<()>, Error> + 'static,
    BIQ: FnOnce(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TFullBackend<Block>>, &Configuration, Option<TelemetryHandle>, &TaskManager, Arc<FrontierBackend<Block>>, BlockNumber) -> Result<DefaultImportQueue<Block, TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, Error>,
    BIC: FnOnce(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TFullBackend<Block>>, Option<&Registry>, Option<TelemetryHandle>, &TaskManager, Arc<dyn RelayChainInterface>, Arc<FullPool<Block, TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>>, Arc<NetworkService<Block, Hash>>, SyncCryptoStorePtr, bool) -> Result<Box<dyn ParachainConsensus<Block>>, Error>,

Start a node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.