Module centrifuge_chain::rpc::evm

Structs

Deps

Extra dependencies for Ethereum compatibility.

EthBlockDataCacheTask

Manage LRU caches for block data and their transaction statuses. These are large and take a lot of time to fetch from the database. Storing them in an LRU cache will allow to reduce database accesses when many subsequent requests are related to the same blocks.

OverrideHandle

RuntimeApiStorageOverride

A wrapper type for the Runtime API. This type implements StorageOverride, so it can be used when calling the runtime API is desired but a dyn StorageOverride is required.

SchemaV1Override

An override for runtimes that use Schema V1

SchemaV2Override

An override for runtimes that use Schema V2

SchemaV3Override

An override for runtimes that use Schema V3

Traits

StorageOverride

Something that can fetch Ethereum-related data. This trait is quite similar to the runtime API, and indeed oe implementation of it uses the runtime API. Having this trait is useful because it allows optimized implementations that fetch data from a State Backend with some assumptions about pallet-ethereum’s storage schema. Using such an optimized implementation avoids spawning a runtime and the overhead associated with it.

Functions

create

Instantiate Ethereum-compatible RPC extensions.

overrides_handle

Type Definitions

FeeHistoryCache

FeeHistoryCacheLimit

Maximum fee history cache size.

FilterPool

On-memory stored filters created through the eth_newFilter RPC.